fix: rewrite all existing modules on typescript.
