+++ index.ts
@@ -1,6 +1,5 @@
-import { httpServer } from "./src/http_server/index.js";
-import { wsServer } from "./src/ws_server/index.js";
-
+import { httpServer } from "./src/http_server";
+import { wsServer } from "./src/ws_server";
 
 const WS_PORT = 8080;
 wsServer.listen(WS_PORT);
@@ -8,4 +7,3 @@ wsServer.listen(WS_PORT);
 const HTTP_PORT = 8181;
 console.log(`Start static http server on the ${HTTP_PORT} port!`);
 httpServer.listen(HTTP_PORT);
-
