+++ src/ws_server/handlers/index.ts
@@ -1,0 +1,1 @@
+export * from "./mouse-up-command-handler";
