+++ src/ws_server/model/command-handler.ts
@@ -1,0 +1,10 @@
+export abstract class CommandHandler {
+  protected next: CommandHandler;
+
+  abstract handle(command: string): Promise<void>;
+
+  setNext(next: CommandHandler) {
+    this.next = next;
+    return this;
+  }
+}
