+++ src/ws_server/model/index.ts
@@ -1,0 +1,1 @@
+export * from "./command-handler";
